refactor: extract SmartScraper adapters into adapters/ package (#38)
SmartScraper reduced from 1,197 to 524 lines. Six domain-specific
adapters extracted into src/justscrape/adapters/:

- reddit.py — RedditAdapter (JSON API scraping)
- stackexchange.py — StackExchangeAdapter (API + StackPrinter fallback)
- devto.py — DevToAdapter (public API)
- github.py — GitHubDiscussionsAdapter (HTML scraping)
- base.py — SourceAdapter ABC + shared _safe_get utility
- __init__.py — get_adapter(url) router

SmartScraper._try_source_adapter() now delegates to get_adapter().
Public interface unchanged. Test patches updated to new _safe_get path.
